Split WGShip.cs into dedicated files for each group of components
