+++ DataConverter/Services/DataConverterService.cs
@@ -11,6 +11,7 @@ using GameParamsExtractor.WGStructure;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using WowsShipBuilder.GameParamsExtractor.WGStructure;
+using WowsShipBuilder.GameParamsExtractor.WGStructure.Ship;
 
 namespace DataConverter.Services;
