+++ WowsShipBuilder.GameParamsExtractor/WGStructure/Ship/WgFireControl.cs
@@ -1,0 +1,10 @@
+﻿// ReSharper disable UnusedAutoPropertyAccessor.Global
+// ReSharper disable CollectionNeverUpdated.Global
+namespace WowsShipBuilder.GameParamsExtractor.WGStructure.Ship;
+
+public class WgFireControl : WgArmamentModule
+{
+    public decimal MaxDistCoef { get; set; }
+
+    public decimal SigmaCountCoef { get; set; }
+}
