+++ WowsShipBuilder.GameParamsExtractor/WGStructure/WgStructureHelper.cs
@@ -1,6 +1,7 @@
 using GameParamsExtractor.WGStructure;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using WowsShipBuilder.GameParamsExtractor.WGStructure.Ship;
 
 namespace WowsShipBuilder.GameParamsExtractor.WGStructure;
 
@@ -18,11 +19,11 @@ public static class WgStructureHelper
         }
     }
 
-    public static Dictionary<string, AaAura> FindAaAuras(this Dictionary<string, JToken> dictionary)
+    public static Dictionary<string, WgAaAura> FindAaAuras(this Dictionary<string, JToken> dictionary)
     {
         return dictionary.Where(entry => entry.Value.Type == JTokenType.Object)
             .Where(entry => entry.Value["hitChance"] != null || entry.Value["HitChance"] != null)
-            .Select(entry => new KeyValuePair<string, AaAura?>(entry.Key, entry.Value.ToObjectOrNull<AaAura>()))
+            .Select(entry => new KeyValuePair<string, WgAaAura?>(entry.Key, entry.Value.ToObjectOrNull<WgAaAura>()))
             .Where(entry => entry.Value is not null)
             .ToDictionary(entry => entry.Key, entry => entry.Value!);
     }
